financial_risk_scanner/config.py
import os
from dotenv import load_dotenv

load_dotenv()

GROQ_API_KEY = os.getenv("GROQ_API_KEY")
MODEL_NAME = "llama-3.3-70b-versatile"
TEMPERATURE = 0.2

N8N_WEBHOOK_URL = "http://localhost:5678/webhook/risk-alert"