financial_risk_scanner/requirements.txt
streamlit
langchain
langchain-community
chromadb
pypdf
sentence-transformers
groq
requests
pydantic
python-dotenv
matplotlib